+++ _tests/test/core/linker/integration/misc_test.dart
@@ -128,16 +128,16 @@ class ChildCompUsingService {
 class DynamicViewport {
   Future<dynamic> done;
 
-  // ignore: deprecated_member_use
-  DynamicViewport(ViewContainerRef vc, ComponentResolver compiler) {
+  DynamicViewport(ViewContainerRef vc) {
     final myService = new MyService()..greeting = 'dynamic greet';
     final injector = new Injector.map({
       MyService: myService,
     }, vc.injector);
-    // ignore: deprecated_member_use
-    done = compiler.resolveComponent(ChildCompUsingService).then(
-        (componentFactory) =>
-            vc.createComponent(componentFactory, 0, injector));
+    final factoryFuture = new Future.value(
+      ng_generated.ChildCompUsingServiceNgFactory,
+    );
+    done = factoryFuture.then((componentFactory) =>
+        vc.createComponent(componentFactory, 0, injector));
   }
 }